+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2.1.1 under development
 
 - Bug #123: Explicitly marking parameters as nullable (@Tigrov)
-- Bug #123: Fix `ContextValueExtractor::extract()` with empty key in PHP 8.4 (@Tigrov)
+- Enh #123, #124: Minor refactor internal class `ContextValueExtractor` (@Tigrov, @vjik)
 
 ## 2.1.0 July 03, 2024